+++ examples/advanced/src/plugins/render.jsx
@@ -3,15 +3,15 @@
  * Render changes to the screen
  */
 
-import Layout    from '../views/layout'
-import Microcosm from 'Microcosm'
-import React     from 'react'
+import Layout     from '../views/layout'
+import Microscope from 'addons/microscope'
+import React      from 'react'
 
 export function register (app, { el }, next) {
   let ui = (
-    <Microcosm instance={ app }>
+    <Microscope instance={ app }>
       <Layout app={ app } />
-    </Microcosm>
+    </Microscope>
   )
 
   React.render(ui, el)